requirements.txt
anthropic>=0.18.0
openai>=1.12.0
httpx>=0.27.0
rich>=13.7.0
jinja2>=3.1.0
pyyaml>=6.0
tiktoken>=0.5.0
